mfd: rk808: turn MFD_RK8XX into a tristate
Prior to commit c20e8c5b1203 ("mfd: rk808: Split into core and i2c")
it was possible to build rk808 as a module and have CONFIG_MFD_CORE=m too.
After that the intermediary CONFIG_MFD_RK8XX was added as a boolean,
forcing mfd_core to be compiled in whenever rk808 was enabled,
even if it was still built as a module.

Signed-off-by: Ivan Babrou <ivan@cloudflare.com>
Fixes: c20e8c5b1203 ("mfd: rk808: Split into core and i2c")